Update docker-compose.yml
Ok, got the compose version of this to run and drop a JSON file locally. Still need a cleaner start since the linkprowler image needs to restart a few times until Selenium is accepting connections. Looking at rolling in "wait-for-it.sh" from some articles I saw while trying to solve the startup issues.